change webpack config file name (#6)
* add npm packages

This adds all packages that are absolutely necessary and excludes anything for testing, which we can add at a later point.

* remove create-react-app dependency and add webpack files

* update react-router-dom

* convert to TS

I did not have to add any type information, so the source code is left as is. TypeScript can infer types. However, it is good practice to add type information to function parameters and return values. We could later add a lint rule to require types.

* add types

* change webpack config file name